requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
cycler==0.10.0
dj-database-url==0.5.0
Django==1.11.7
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.2
pandas==0.21.0
Pillow==5.4.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pycparser==2.19
PyMySQL==0.9.2
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2017.3
scipy==1.2.0
six==1.11.0
whitenoise==4.1.2
